requirements.txt
numpy==1.24.4
pandas==2.0.3
Pillow==10.1.0
scikit-learn==1.3.2
streamlit==1.28.2
streamlit_option_menu==0.3.6
